requirements.txt
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
seaborn==0.11.2
plotly==5.3.1
scikit-learn==0.24.2
xgboost==1.4.2
jupyter==1.0.0
requests==2.26.0
beautifulsoup4==4.10.0
scipy==1.7.1
statsmodels==0.13.0
python-dotenv==0.19.0